--- ClientApp/src/components/admin/dashboard-all/NEW/ScrollerComponent.tsx
@@ -1,23 +1,0 @@
-import React, { useCallback } from "react";
-//import { OverlayScrollbarsComponent } from "overlayscrollbars-react";
-
-/*
-const CustomScrollbar = React.forwardRef(({ children, className, style }, ref) => {
-  const refSetter = useCallback(
-    (scrollbarsRef) => {
-      if (scrollbarsRef) {
-        ref.current = scrollbarsRef.osInstance().getElements().viewport;
-      }
-    },
-    [ref]
-  );
-
-  return (
-    <OverlayScrollbarsComponent ref={refSetter} className={className} style={style}>
-      {children}
-    </OverlayScrollbarsComponent>
-  );
-});
-
-export default CustomScrollbar;
-*/
