+++ ClientApp/src/components/admin/dashboard-all/RecipientDataTable.tsx
@@ -1,9 +1,9 @@
 import { Box, Divider, Grid, MenuItem, Select, TextField, Typography } from "@material-ui/core";
 import React, { useState } from "react";
-import useStyles from "../Styles";
-import { GiverType } from "../../../shared/Types";
+import useStyles from "./Styles";
+import { RecipientType } from "../../shared/Types";
 import { Virtuoso } from "react-virtuoso";
-import GiverDataCard from "./GiverDataCard";
+import RecipientDataCard from "./RecipientDataCard";
 import { Search } from "@material-ui/icons";
 import { CONNECTION_COLORS } from "common/constants/ConnectionColors";
 import { FAMILY_SIZES } from "common/constants/FamilySizes";
@@ -11,20 +11,20 @@ import PeopleOutlineIcon from "@material-ui/icons/PeopleOutline";
 import LinkIcon from "@material-ui/icons/Link";
 
 type Props = {
-  giverData: GiverType[];
-  selectedGiverIndex: number;
-  setSelectedGiver: (giver: GiverType) => void;
-  setSelectedGiverIndex: (idx: number) => void;
+  recipientData: RecipientType[];
+  selectedRecipientIndex: number;
+  setSelectedRecipient: (giver: RecipientType) => void;
+  setSelectedRecipientIndex: (idx: number) => void;
   refreshData: () => void;
   accessToken: string;
   resetSelections: () => void;
 };
 
-const GiverDataTable: React.FC<Props> = ({
-  giverData,
-  selectedGiverIndex,
-  setSelectedGiver,
-  setSelectedGiverIndex,
+const RecipientDataTable: React.FC<Props> = ({
+  recipientData,
+  selectedRecipientIndex,
+  setSelectedRecipient,
+  setSelectedRecipientIndex,
   refreshData,
   accessToken,
   resetSelections,
@@ -34,22 +34,29 @@ const GiverDataTable: React.FC<Props> = ({
   const [familySize, setFamilySize] = useState(-1);
   const [color, setColor] = useState("alle");
 
-  const searchFilter = (input: GiverType[] | []) => {
+  const searchFilter = (input: RecipientType[]) => {
     return input.filter(
       (input) =>
-        input.fullName?.toLowerCase().indexOf(query.toLowerCase()) > -1 ||
-        input.email?.toLowerCase().indexOf(query.toLowerCase()) > -1 ||
-        input.phoneNumber?.toLowerCase().indexOf(query.toLowerCase()) > -1
+        input.contactEmail?.toLowerCase().indexOf(query.toLowerCase()) > -1 ||
+        input.contactFullName?.toLowerCase().indexOf(query.toLowerCase()) > -1 ||
+        input.contactPhoneNumber?.toLowerCase().indexOf(query.toLowerCase()) > -1 ||
+        input.institution?.toLowerCase().indexOf(query.toLowerCase()) > -1 ||
+        input.referenceId?.toLowerCase().indexOf(query.toLowerCase()) > -1 ||
+        input.familyId.toString()?.toLowerCase().indexOf(query.toLowerCase()) > -1
     );
   };
-  const familyFilter = (input: GiverType[] | []) => {
+  const rounUpFamilyMembers = (famSize: number) => {
+    const matchFamilySize = FAMILY_SIZES.find((famObj) => famObj.value >= famSize)?.value;
+    return matchFamilySize;
+  };
+  const familyFilter = (input: RecipientType[]) => {
     if (familySize == -1) {
       return input; // Unfiltered
     } else {
-      return input.filter((input) => input.maxReceivers == familySize);
+      return input.filter((input) => rounUpFamilyMembers(input.familyMembers.length) == familySize);
     }
   };
-  const colorFilter = (input: GiverType[] | []) => {
+  const colorFilter = (input: RecipientType[]) => {
     if (color == "Rød") {
       return input.filter((input) => !input.isSuggestedMatch && !input.hasConfirmedMatch);
     } else if (color == "Gul") {
@@ -61,7 +68,7 @@ const GiverDataTable: React.FC<Props> = ({
     }
   };
 
-  const allFilters = (input: GiverType[] | []) => {
+  const allFilters = (input: RecipientType[]) => {
     let filteredInput = input;
     filteredInput = searchFilter(filteredInput);
     filteredInput = familyFilter(filteredInput);
@@ -80,15 +87,9 @@ const GiverDataTable: React.FC<Props> = ({
   return (
     <>
       <Typography align="center" variant="h5">
-        Givere
+        Familier
       </Typography>
-      <Grid
-        container
-        direction="row"
-        alignItems="center"
-        className={classes.tableHeadingSpace}
-        spacing={1}
-      >
+      <Grid container direction="row" alignItems="center" className={classes.tableHeadingSpace}>
         <Grid item xs={4}>
           <Box className={classes.gridBoxCenter}>
             <Search />
@@ -147,15 +148,14 @@ const GiverDataTable: React.FC<Props> = ({
       <Divider />
 
       <Virtuoso
-        totalCount={allFilters(giverData).length}
-        style={{ marginBottom: "2em" }}
+        totalCount={allFilters(recipientData).length}
         itemContent={(index) => (
-          <GiverDataCard
-            giverData={allFilters(giverData)[index]}
-            giverIndex={index}
-            selectedGiverIndex={selectedGiverIndex}
-            setSelectedGiver={() => setSelectedGiver(allFilters(giverData)[index])}
-            setSelectedGiverIndex={() => setSelectedGiverIndex(index)}
+          <RecipientDataCard
+            recipientData={allFilters(recipientData)[index]}
+            recipientIndex={index}
+            selectedRecipientIndex={selectedRecipientIndex}
+            setSelectedRecipient={() => setSelectedRecipient(allFilters(recipientData)[index])}
+            setSelectedRecipientIndex={() => setSelectedRecipientIndex(index)}
             refreshData={refreshData}
             accessToken={accessToken}
             resetSelections={resetSelections}
@@ -165,4 +165,4 @@ const GiverDataTable: React.FC<Props> = ({
     </>
   );
 };
-export default React.memo(GiverDataTable);
+export default React.memo(RecipientDataTable);
